[openssl] Consistently use OPENSSL_VERSION instead of just VERSION
The inner CMakeLists.txt used to reference ${VERSION} everywhere except
as the working directory for patching, where ${OPENSSL_VERSION} was used
and the patches did not get applied because of that.

This changes it to use ${OPENSSL_VERSION} everywhere instead, consistent
with the name of this variable in the parent portfile.cmake.