[geotiff] Hack: include find_package(PROJ4) in config cmake
This feels like a bit of a hack. If a consumer of geotiff does not
call find_package(PROJ4), CMake won't be able to resolve the PROJ4
dependency. This will result in "proj.lib" instead of its actual full
path to be added as a dependency (when using CMake to build a project),
causing an error during the linking phase. When find_package(PROJ4) is
called, CMake will add the correct paths to projd.lib and proj.lib.