Fix Win10SDK detection during Bootstrap
When multiple Win10 SDK's are installed, like Insider previews,
depending on the SDK installation, it may not contain the um headers.
There is already a loop for trying to find the newest Win10 SDK
that has full headers, but it's returning at the first failure,
rather than continuing the loop to the next iteration.

Also, use Join-Path instead of manual path appending, to normalize
the extra \ at the end of the sdk root value.  In some cases paths
with double backslashes can be interpreted as a fallback syntax for
UNC paths.  Join-Path normalizes the path appropriately.