curl: update to v7.57.0
Due to changes in curl's CMake support, where it now installs .cmake
files, we now have to make sure that they are installed into the correct
directory (and not duplicated into the debug/share/ directory, either).

Also, a change in the context of the 2nd patch required an update of
said patch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>