[libspatialite] fix compile error on non-latin env
It has literals with latin chars in a C source.
Unfortunately VC++ recognize source file without BOM as in OEM code not in UTF-8.
It cause a strange behavior of VC++ compiler, and fails with
"epsg_inlined_prussian.c(685): error C2001: newline in constant."
It happens on non-latin Windows such as Japanese(CP932).

It also work when changing Windows configuration
"system locale for non-unicode applications" to latin one such as English(US)(CP432),
but it affects all applications and users in Windows system.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>