Require the version of g++ to be at least 6
On Ubuntu 16.04, old versions of g++ such as 4.9 are available. Other
Linux distributions may also ship with g++ versions older than 5. With
the previous test, the bootstrap script was not failing fast and instead
was starting to compile vcpkg before it encountered the unsupported flag
-std=c++1z. The error message was not very explicit and that was making
it difficult to the developer to understand what the problem was. However,
when using g++ 5, an clear error message was printed to help the user.
This commit shows the helpful error message anytime g++ is older than
version 6.