[qt5] Regression fix: Move debug/lib/Qt5Bootstrap.* elsewhere
This commit fixes part of a previous issue:
https://github.com/Microsoft/vcpkg/issues/1442

In this older issue, Qt5Bootstrap.lib was originally placed in
debug/lib, causing it to be automatically linked in Visual Studio
for debug builds. This caused linking errors, since Qt5Bootstrap.lib
is linked against a different runtime (as detailed in the issue).

The solution back then, as implemented in PR
  https://github.com/Microsoft/vcpkg/pull/1847
, was to place this file elsewhere.

The same approach is taken here in this commit. However, the Qt5 port
is nowadays modular and complicates matters. The qt5-base package
produces the problematic .lib file and should be responsible for moving
the file elsewhere, yet at the same time the qt5-tools package
depends on this .lib file.

Ideally, the qt5-tools port should be patched such that it would find
Qt5Bootstrap.lib in its new location. Unfortunately, the path to
Qt5Bootstrap.lib appears in a Makefile that is generated by qmake based
on a *.pro and a qt.conf file, a process which I do not yet fully
understand.

As an alternative, I opted for temporarily copying the Qt5Bootstrap files
to their original location before building the qt5-tools package and
clean up afterwards.