breakpad: Enable building the libbreakpad_client library
That's the library that allows creating and uploading .dmp files
when a program crashes.

Use the same flags and setup as for libbreakpad, just note that
this commit changes the installation folder for headers which
was previously incorrect:
The libbreakpad headers were installed in include/google_breakpad,
but the correct path is include/breakpad/google_breakpad,
and client headers go into include/breakpad/client/windows.