[GraphicsMagick] bug-fix: image files won't load
GM (GraphicsMagick) was failing to load image files, notably PNG or JPEG
images, as it was being compiled with support for 'GraphicsMagick
Modules'.  These are files with names specific to GM, and of the format,
'IM_*.dll'.  vcpkg's install process was not setting these up.  This
patch makes sure that when GM is built, it embed's GM's own
image-loading code into graphicsmagick.dll.