Add documentation about Abseil package name and namespace
Putting this at the top of the CMakeLists.txt file would help keep users from having to read through the CMAkeLists.txt file looking for the correct information.  If there is a better place for this, please let me know.  I could also submit a little example project using this package if there's a place for it.