Fix build of OpenSSL on Android (#4427)
When passing on CMAKE_<LANG>_COMPILER_TARGET and
CMAKE_<LANG>_COMPILER_EXTERNAL_TOOLCHAIN it is important that the
selection of those via CMAKE_C_COMPILE_OPTIONS_TARGET/EXTERNAL_TOOLCHAIN
is done without a space inbetween. For example COMPILE_OPTIONS_TARGET is
with clang a simple --target= and so the selected target must follow
without a space, so that we get --target=foo-bar instead of --target=
foo bar, which the compiler does not parse as intended.