Add initial support for RapidCheck
Create a portfile for the C++ property-based testing library RapidCheck.
This port is relatively straight forward but there are a couple of caveats:
 * The current port uses a branch from a fork that I modified as the CMake installation logic had some problems.
   The library was almost exclusively used by being added as a subfolder so the CMake export machinery was probably not well tested.
 * The changes are going to get merged into the main repository eventually and then it would be possible to change where the source
   code is downloaded.

I am unaware when the main RapidCheck library maintainer will accept my pull request, so it is better to have this library in VCPKG
sooner, rather than later.