[wxwidgets] remove include/msvc directory
The only thing in include/msvc is include/msvc/wx/setup.h which is a
"wrapper" around the real setup.h. The wrapper setup.h is hard-coded to
include the real setup.h from lib but since the real setup.h is now in
include/wx the wrapper has become useless.