Add a more correct port for libass
This port currently only works on Windows because the dependencies of
the port (fribidi in particular) don't build on non-Windows platforms.

I made patch modifications to the source code because the Visual C++ C
compiler does not allow for non-constant-expression array indicesto be
used to declare a static array.

I've verified that this (1) builds on Windows and (2) works with a
sample test that's included in the libass github repo.