[libgeotiff] Remove FindPROJ4.cmake from src
The GeoTIFF library ships with its own FindPROJ4.cmake module. This file
interferes with the cmake files that are installed by vcpkg for the PROJ4
package. As a result, the debug version of libgeotiff gets incorrectly
linked to the release version of PROJ4. By removing FindPROJ4.cmake from
the source directory, the vcpkg's .cmake files are used instead,
allowing CMake to find the debug version of PROJ4 when building the
debug version of libgeotiff.