Force static linkage for amqpcpp
The documentation for this library says that shared linkage on Windows
is unsupported. Furthermore, the build currently errors out on Windows
because shared DLLs are not placed correctly.

vcpkg only supports static linkage on Linux and macOS, so I'm
categorically making amqpcpp build with static linkage on all platforms.
This unifies the platforms and gives consistent behavior