[json-spirit] Add initial port (#5157)
* [json-spirit] Add initial port

* Modify PR to using pn85, enable some features

json-spirit is a weird case because the main repo that package managers
tend to use appears to not be actively maintained. Nevertheless, vcpkg
tries to use the "official" source of a repository if at all possible.

In order to do this, I've made the following changes to the PR:
- I've changed the upstream REPO, REF, and SHA512.
- I've passed options to disable building the demos and tests
- I've made the project build only as a static library.

The reason for this is that vcpkg does not support dynamical linkin.
Another reason is that the library as it is in the repo does not have
all the definitions correctly annotated.smlee-hdactech has made a PR
against the major upstream repo to fix this, butwe are going to defer
to the main repo until that change is accepted.