Add X-Plane SDK support (#5408)
* Add X-Plane SDK support

The x-plane SDK website claims that the SDK is only
compatible with x64. The PR that adds support for this
fails in x86, so I'm disallow x86 builds to align with
the x-plane documentation.

The x-plane SDK website's examples expect that the sdk
headers are at the root of the include path. I'm adjusting
this port to align with those samples, as we want examples
to "just work" as much as possible.