Install CGAL as header-only (#4975)
* Update to install CGAL as header-only

I think the issue #2471 would be completely solved, if vcpkg installed CGAL as header-only. That way, the installed `CGALConfig.cmake` would search for its dependencies, instead of hard-coding their path in the file.

* Bump the version

* Fix the cleaning

Once CGAL is header-only, no binary library is created, and the
`debug/` directory got empty.

This commit fixes the cleaning: now the whole `debug/` directory is
removed.