[icu] Prevent stale MSYS gpg-agent.exe process blocking command control (#6407)
* Prevent stale MSYS gpg-agent.exe process blocking command control

This commit fixes:
https://github.com/microsoft/vcpkg/issues/5476

The issue is that CI environments such as Appveyor's VS2017 image will wait for all processes to complete. If a stale process resides as a result, builds will hang.
There does not appear to be any good reason for gpg-agent.exe to be running once the build of icu has completed.

Without this patch builds of icu4c using CI systems will very likely hang and not in an obvious way.

Is this the _right_ solution to this problem? Probably not but it is one solution. And it degrades gracefully in that the build will not fail if gpg-agent.exe is not running. The gpg-agent.exe will not run again once MSYS has been configured, so to test this patch, a fresh install of vcpkg is required. Open the task manager and before the icu build completes, look for gpg-agent.exe just sitting there for no reason.
Might I suggest that the issue is fixed in vcpkg MSYS instead or as well?

Please don't request further from this commit.

* [icu] Kill MSYS gpg-agent.exe on Windows