[minizip] Make BZip2 an optional feature (#7282)
* [minizip] Update CMakeLists.txt

If the BZip2 package has not been found, ensure that the library does
not assume it is. This would stop any build on targets where bzip2 is
not installed. This can either be the bzip2 package provided by vcpkg
or locally on the system.

Feature:
- Allow optionally to enable bzip2 supprt

* [minizip] Make bzip2 truly optional

Also catch dependency in the corresponding cmake files