[Proposal] Add a new triplet for static libraries with dynamic CRT linkage
Static libraries with dynamic CRT linkage are useful if one wants to build a dynamic library further down the line that links its dependencies as static libraries instead of dynamic ones. For example, consider `libfoo` that depends on `libbar` and `libquux`: if one was to distribute `libfoo.dll`, right now they would have to distribute `libbar.dll` and `libquux.dll` as well. With this triplet, it would be possible to build `libbar.lib` and `libquux.lib` and link them into `libfoo.dll`, which would then be self-sufficient.