[qt5-base] whitespace change to clean CI system
The cached version of qt5 in the CI system was built across multiple Linux VMs.  This has caused the paths in the archived files to be inconsistent.  Making a whitespace change in qt5-base and forcing CI to run on a single machine to see if it cleans out the problem.