Workaround Powershell 6.2/7.0 breaking compatibility with Get-ChildItem
Powershell 6.2/7.0 in infinite wisdom decided to change the value of
what `Get-ChildItem` (and anything dealing with FileInfo/DirectoryInfo)
returns from relative to absolute path.

The only way to work around this for both existing versions and future
versions is to explicitly request the `Name`. Hopefully this doesn't
get changed again.

NOTE: This PS change breaks a lot of scripts, including many
of Microsoft's own first-party scripts. No workarounds were pushed
for any that I could see in the year+ since the issue was first
reported.