Including config file openssl.cnf in installation.
The original cmake script after the build delete the openssl.cnf files,
but if there is no config file in the environment the openssl.exe
warnings and fails in some commands.

This commit instead of deleting one of the openssl.cnf file it renames
(moves) to the installation dir tools/openssl, the same openssl.exe
final destination.