[sol2] Use the single header release
Using the single header release of the library is easier, faster to 'build', and is recommended by the documentation. It also fix the duplicated include folder issue of the previous port : <sol/sol/sol.hpp> can now be replaced with <sol/sol.hpp> as it should be.

Is there any drawback about using the single header release that I am not seing?