[libffi] Add libffiConfigVersion.cmake file
Use write_basic_package_version_file() to create the
libffiConfigVersion.cmake file.
This allows detection of the libffi version using cmake.
An example, where this is relevant is e.g. glib 2.62, with the
following code in meson.build:
  libffi_dep = dependency('libffi', version : '>= 3.0.0',
  fallback : ['libffi', 'ffi_dep'])

When the libffiConfigVersion.cmake file is present, the following output
can be found in the meson-build.log:
  Run-time dependency libffi found: YES 3.1