[faad2] Add initial version of FAAD2
FAAD2 lists its homepage as https://sourceforge.net/projects/faac/, but
it links to https://github.com/knik0/faad2 as the place to send pull
requests. This seems good enough for Debian[0] and Arch[1], and the SF
hosts only a bit older versions, so I used the GitHub project as well.

Note that though the project is named "faad2", the library name is just
"faad"

The embedded patches were all submitted and already merged upstream.

[0] https://packages.debian.org/sid/faad
[1] https://www.archlinux.org/packages/extra/x86_64/faad2/