Prevent python3 build failure (#9014)
* Prevent python3 build failure

If VCPKG_BUILD_TYPE release is set inside a
triplet, then the build will fail because
the port file attempts to deal with debug
files which will never exist. The changes
in this patch allow the build to succeed
if VCPKG_BUILD_TYPE release is stipulated
using a triplet before a build of the
python3 port is initiated.

* python3 version bump

Port clean up and version number
bump.

* libxslt release triplet build fix

This patch fixes a condition where
if a debug build is disabled via
triplet, the build fails and an
empty directory error message
is sent to the console.