[gtk] Bump to gtk+-3.24.10 which is the first version using meson build
- Fixes #7412 by using gtk 3.24.10.

- Possibly fixes #6554, #7827, #8103 as now all the missing assets are
correctly copied into installed/<triplet>/share/ and should be deployed
with the final application.

This simplify the gtk port and correctly generate the missing files.

The source of the problem is with the underlying build system, as these
files were installed by the demo `icon-browser`, but the
gtk-update-icon-cache were executed by the main gtk project, failing if
the demo wasn't built or installed.

- This also fixes #7641 as now the meson build will correctly generate
the missing .pc files.

- This also fixes #9035, #7150 by fixing `vcpkg_configure_meson.cmake`
which fails on Linux/OSX because it indiscriminately try to pass msvc
cl arguments to the cmake c/xx flags. Although this works, it is not
the ideal fix. The correct fix would be to do as suggested in
issue #8271. Also see issue #8630.