[cpprestsdk] Disable websockets by default.
The websockets backend cpprestsdk uses is websocketpp, which was last
committed to in late 2018 and appears defunct. Additionally, the
websockets feature brings an otherwise unnecessary Boost dependency on
Windows which takes a relatively long time to build.

Customers who still want the websockets bits can turn on the optional feature.