ignition-modular-scripts: Migrate from bitbucket to github
As announced in https://community.gazebosim.org/t/important-gazebo-and-ignition-are-going-to-github/533,
the ignition project repositories have been migrated from BitBucket to GitHub.

This commit updates the ignition_modular_library helpers to use GitHub instead of BitBucket,
and also update the hashes of all the ignition ports as apparently the archive generated by GitHub
for releases are slightly different from the one generated by BitBucket.