Use vcpkg_build_make on Linux and Mac, but only compile in release mode.
Its difficult to get the right binary if there are both debug and release builds and `vcpkg_build_make` will compile out of place. Also, the premake makefiles do not support an "install" target.
Instead of reconstructing the logic in `vcpkg_build_make`, we just disable debug builds.
Since premake5 is an executable, this should not be an issue.