protobuf: Do not install malformed CMake configuration files when crosscompiling
Before this commit, the protobuf portfile when crosscompiling was not defining the
protobuf::protoc target in its CMake config files, even if its definition was instead
correctly patched in the port to point to the host version of the tool. Furthermore,
the protobuf::libprotoc library (that is a library that can also be used in
cross-compilation build) was not installed.