[osg] Fix conflict when asio and boost-asio are installed.
When asio and boost-asio are installed, osg's build scripts would choose the standalone version of asio, but uses features specific to boost-asio. This patch forces the port to always use boost-asio, as the original CONTROL file indicated was the intent.

Also fixed transitive dependencies not being declared that just happened to get installed because boost-asio is a dependency.