[vcpkg] Provide $(VcpkgRoot) and $(VcpkgCurrentInstalledDir) for customers.
This is a competing resolution with https://github.com/microsoft/vcpkg/pull/11772

In https://github.com/microsoft/vcpkg/commit/4fb225608532e9fb2fd2f5f1dbe9ec092cdc7c93#diff-f7a49059a86ff0bb925110f33551a677R57 I removed all uses of $(VcpkgRoot) and $(VcpkgCurrentInstalledDir) because sometimes those properties were passed on the command line or from a user project where they are likely to not have a trailing slash. The intent was to always use VcpkgRootSanitized or similar, which was the same as VcpkgRoot but with a trailing slash added if one was missing from the input. (We could not use the same name because msbuild ignores any modifications of properties set on the command line when msbuild is invoked by default)

This change uses TreatAsLocalProperty to do the add-trailing-slash sanitization, without changing the property name, because it turned out folks were using that property in their own build scripts.