Change how vcpkg-configuration works after discussion with team
We had a serious discussion today, and so the way we're going forward
with this is very different -- no longer having a `"configuration"`
field, adding inheritance, etc.

Check the RFC for more info on how this will be done in the future; for
the MVP, all it changes (fundamentally) is where the configuration stuff
gets set -- it's in vcpkg-configuration.json in the manifest root.