[vcpkg_fixup_pkgconfig] Remove detection of individual libraries
Because pkgconfig's recommended approach when using libraries outside pkgconfig is to add them to your .pc file, it is impossible to enumerate the set of external libraries on all possible systems.

We still run `pkgconfig --exists` to ensure that the .pc file is able to be loaded in the current environment.