[libp7client, libp7-baical] Combine identical ports and update to libp7client 5.6.
In a recent CI build, we started getting output like:

Installing package libp7-baical[core]:x86-windows...
The following files are already installed in C:/Dev/vcpkg/installed/x86-windows and are in conflict with libp7-baical:x86-windows

Installed by libp7client:x86-windows
    bin/P7x32.dll
    bin/P7x32.pdb
    debug/bin/P7x32d.dll
    debug/bin/P7x32d.pdb
    include/P7/GTypes.h
    include/P7/P7_Client.h
    include/P7/P7_Cproxy.h
    include/P7/P7_Extensions.h
    include/P7/P7_Telemetry.h
    include/P7/P7_Trace.h

It turns out that these ports are identical, they are just different versions of the same library. The libp7-baical version supported more platforms and was simpler, so took that one and updated it to a current version.