Revert "GH Actions: use Azure Artifacts binary cache"
This reverts commit 20d349d07952c1775bf0baa3908ee6834a076641.

The most recent build failed to push any packages to Azure
Artifacts. I think that is because there is not enough space left.
Even though I deleted some old packages, Azure Artifacts still
counts the space as used for a while. It may be better to just
cache the builds on GitHub Actions. Now that we are not planning
to use the Azure Artifacts cache for Mixxx builds, there isn't
much a point using Azure Artifacts for the vcpkg environment
builds.