Fix CURL filesystem deadlock with corporate proxy
The CURL call to the metrics  endpoint doesn't have a timeout. When executing this behind a corporate proxy that blocks this access the filesystem is in a deadlock until curl is killed. By adding a 200ms timeout we give ample time for CURL to finish its request, while also not stopping usage inside local network.