[triplets][wasm32-emscripten] Find emscripten installation based on PATH
Not all emscripten users are emsdk users so `$EMSDK` won't always be
present.  However its resonable to assume that all emscripten users
(including emsdk users) will have `emcc` (the esmcripten compiler) in
their `$PATH`

Keep the old method of locating emscripten (which is fragile since it
depends on the internal layout of emsdk) around in case there are any
users that set `$EMSDK` themselves without adding emcc to the `$PATH`.
This is unlikely since emsdk, when used in the normal way, takes care
of adding the emscripten directory to `$PATH`.