[QtWebEngine] - Fix build failure
This commit fixes an issue
where the build fails on
X86_64 Windows 10 Pro
with a message similar to:
'PATCHES no such argument'
printed to the console output.

It also makes a few small
changes for sanity.
Build trees for QtWebEngine
come in at between 40GB-50GB
and the minimum recommended
RAM size for building
Chromium according to
google is 8GB.
The port shouldn't be
building debug by
default due to it's
size and the fact that
you likely only want
debug if you are a
Qt developer or
are working on vcpkg.
The options "" is
intentionally left
blank due to the fact
that it allows the string
-webengine-proprietary-codecs
to be easily be added in case
you have licences for that
kind of thing.
Without that option enabled,
many html5 videos do not
work in QtWebEngine.
Maybe adding a configuration
switch for proprietary-codecs
is something to think about
for the future?

Enjoy!