[harfbuzz] work around upstream SameFileError (#16621)
During the configuration of harfbuzz several Python scripts are invoked
that are supposed to copy some files. For some reason however the
scripts end up being instructed to copy files to themselves at which
point a SameFileError is risen, causing the build to fail.

This is a workaround for the underlaying issue that is deemed to be
usptream. The upstream issues are
- https://github.com/mesonbuild/meson/issues/8375
- https://github.com/harfbuzz/harfbuzz/issues/2870

The workaround is to add a patch that makes sure these scripts check
whether source and target file are the same and only invoking the copy
action if they are not.

Fixes #16262