We've added a couple of functions to allow for better handling of situations described in here:https://www.reddit.com/r/gridnetproject/comments/md2a2m/denial_of_servicedos_attacks_weve_been_facing_and/
The aim was to be able to abort a connection possibly as soon as the incoming connection had been accepted on the listening TCP socket. Thus we wanted to approach the usual way that to us would seem like

1)create ssh_bind

2)initialize ssh_bind_callbacks_struct with ssh_callbacks_init()

3)pass a pointer to a function with a footprint of void callbackConn (ssh_bind sshbind, void* userdata) to incoming_connection of ssh_bind_callbacks_struct which just got initialized (0ed out is seems)

call ssh_bind_set_callbacks() and with that set-up we would have expected for the callbackConn() function to fire.

It never happens. 'normally' we would have accepted the connection in such a function.

The usual accept_bind_accept works but as the practice shows, peers are able to exploit the timeout set (we could change the default timeout of infinity which as practice shows peers might exploit but not completing one of the handshake steps they participate in).
So with accept_bind_accept IF a timeout is set evil peers might still block the server-processing queue up to the timeout value; and if method like-keyboard interactive is enabled we cannot set the timeout to say 2sec as user wouldnt be able to type in credentials at all and if we set to to 10 seconds then the malicious attacker can block processing for all users for up to 10 seconds. By the way; when we disabled keyboard-interactive authentication, the keyboard interactive thing it the first to show up when connecting to server. ssh_bind_listen indeed doesn't seem to accomplish anything more than 'listening', no event processing logic seems to be invoked. ssh_bind_poll_callback never gets called.

We've added a couple of modifications that make the appropariate events fire and also making it possible to abort conneciton on sight without any reply at all.