[brotli] Fold common into dec and enc archives for static builds.
This change allows downstream consumers of the archives to not worry
about ordering the link correctly when using this package, which can be
problematic when --as-needed is passed to the linker and the order isn't
otherwise known.

Keeping the brotlicommon-static archive as is allows for probing of the
brotli libraries to remain symmetric between dynamic and static links.

Fixes static linking issue discussed in
https://github.com/microsoft/vcpkg/issues/16991 which is blocking
rust-sdl2 CI dependent upon vcpkg discussed here
https://github.com/Rust-SDL2/rust-sdl2/pull/1085 .